support for remote actions

+++ scripts/dev.js
@@ -44,59 +44,66 @@ class ActionServer extends CNAScript {
 
     // dev env is needed to generate local actions
     // process.env['NODE_ENV'] = process.env['NODE_ENV'] || 'development'
+    let owStack
+    if (!this.config.actions.remote) {
+      // 1. make sure we have the local binary
+      if (!(await fs.exists(OW_JAR_FILE))) {
+        this.emit('progress', `Could not find ${OW_JAR_FILE}, downloading it from ${OW_JAR_URL}, this might take a while ...`)
+        const content = await request({ url: OW_JAR_URL, followAllRedirects: true, encoding: 'binary' })
+        await fs.writeFile(OW_JAR_FILE, content, 'binary')
+        this.emit('progress', `Successfully downloaded ${OW_JAR_FILE}`)
+      }
 
-    // 1. make sure we have the local binary
-    if (!(await fs.exists(OW_JAR_FILE))) {
-      this.emit('progress', `Could not find ${OW_JAR_FILE}, downloading it from ${OW_JAR_URL}, this might take a while ...`)
-      const content = await request({ url: OW_JAR_URL, followAllRedirects: true, encoding: 'binary' })
-      await fs.writeFile(OW_JAR_FILE, content, 'binary')
-      this.emit('progress', `Successfully downloaded ${OW_JAR_FILE}`)
+      // 2. start the local ow stack
+      this.emit('progress', `starting local OpenWhisk stack..`)
+      owStack = execa('java', ['-jar', '-Dwhisk.concurrency-limit.max=10', OW_JAR_FILE])
+      const logStream = fs.createWriteStream(OW_LOG_FILE) // todo formalize logs in config folder
+      owStack.stdout.pipe(logStream) // todo not showing cleanup logs.. shuting down to early
+      owStack.stderr.pipe(process.stderr) // todo error on stderr ?
+      await waitFor(7000)
+
+      // 3. change the .env
+      if (!(await fs.exists(DOTENV_SAVE))) { await fs.move('.env', DOTENV_SAVE) }
+
+      // Only override needed env vars and preserve other vars in .env
+      const env = dotenv.parse(await fs.readFile(DOTENV_SAVE))
+      // todo don't harcode port
+      env['AIO_RUNTIME_APIHOST'] = 'http://localhost:3233'
+      env['AIO_RUNTIME_AUTH'] = '23bc46b1-71f6-4ed5-8c54-816aa4f8c502:123zO3xZCLrMN6v2BKK1dXYFpXlPkccOFqm12CdAsMgRU4VrNZ9lyGVCGuMDGIwP'
+      env['AIO_RUNTIME_NAMESPACE'] = 'guest'
+      delete env['AIO__RUNTIME_AUTH']
+      delete env['AIO__RUNTIME_NAMESPACE']
+      delete env['AIO__RUNTIME_APIHOST']
+      const envContent = Object.keys(env).reduce((content, k) => content + `${k}=${env[k]}\n`, '')
+
+      await fs.writeFile('.env', envContent)
+      this.emit('progress', `saved .env to ${DOTENV_SAVE}`)
+      this.emit('progress', 'set guest credentials in .env')
+
+      // 4. build and deploy actions // todo support live reloading ? or just doc redeploy
+      this.emit('progress', `redeploying actions to local environment..`)
+      // 4.1 update config -- need to manually reload env vars
+      // hack we need to manually reload env vars, as dotenv is not reloaded
+      // see https://github.com/adobe/aio-cli-config/issues/2
+      // todo generalize this and move to config
+      // this would need to save env vars set outside of .env
+      Object.keys(process.env).forEach(k => { if (k.startsWith('AIO')) delete process.env[k] })
+      dotenv.config() // reload new dotenv
+      const newConfig = require('../lib/config-loader')()
+      // 4.2 do build and deploy to local ow stack
+      await (new BuildActions(newConfig)).run()
+      await (new DeployActions(newConfig)).run()
+
+      // 5. inject new action urls into UI
+      await utils.writeConfig(newConfig.web.injectedConfig, newConfig.actions.urls)
+    } else {
+      // todo deploy
+      // todo live redeploy?
+      this.emit('progress', `using remote actions`)
+      await utils.writeConfig(this.config.web.injectedConfig, this.config.actions.urls)
     }
 
-    // 2. start the local ow stack
-    this.emit('progress', `starting local OpenWhisk stack..`)
-    const owStack = execa('java', ['-jar', OW_JAR_FILE])
-    const logStream = fs.createWriteStream(OW_LOG_FILE) // todo formalize logs in config folder
-    owStack.stdout.pipe(logStream) // todo not showing cleanup logs.. shuting down to early
-    owStack.stderr.pipe(process.stderr) // todo error on stderr ?
-    await waitFor(7000)
-
-    // 3. change the .env
-    if (!(await fs.exists(DOTENV_SAVE))) { await fs.move('.env', DOTENV_SAVE) }
-
-    // Only override needed env vars and preserve other vars in .env
-    const env = dotenv.parse(await fs.readFile(DOTENV_SAVE))
-    // todo don't harcode port
-    env['AIO_RUNTIME_APIHOST'] = 'http://localhost:3233'
-    env['AIO_RUNTIME_AUTH'] = '23bc46b1-71f6-4ed5-8c54-816aa4f8c502:123zO3xZCLrMN6v2BKK1dXYFpXlPkccOFqm12CdAsMgRU4VrNZ9lyGVCGuMDGIwP'
-    env['AIO_RUNTIME_NAMESPACE'] = 'guest'
-    delete env['AIO__RUNTIME_AUTH']
-    delete env['AIO__RUNTIME_NAMESPACE']
-    delete env['AIO__RUNTIME_APIHOST']
-    const envContent = Object.keys(env).reduce((content, k) => content + `${k}=${env[k]}\n`, '')
-
-    await fs.writeFile('.env', envContent)
-    this.emit('progress', `saved .env to ${DOTENV_SAVE}`)
-    this.emit('progress', 'set guest credentials in .env')
-
-    // 4. build and deploy actions // todo support live reloading ? or just doc redeploy
-    this.emit('progress', `redeploying actions to local environment..`)
-    // 4.1 update config -- need to manually reload env vars
-    // hack we need to manually reload env vars, as dotenv is not reloaded
-    // see https://github.com/adobe/aio-cli-config/issues/2
-    // todo generalize this and move to config
-    // this would need to save env vars set outside of .env
-    Object.keys(process.env).forEach(k => { if (k.startsWith('AIO')) delete process.env[k] })
-    dotenv.config() // reload new dotenv
-    const newConfig = require('../lib/config-loader')()
-    // 4.2 do build and deploy to local ow stack
-    await (new BuildActions(newConfig)).run()
-    await (new DeployActions(newConfig)).run()
-
     this.emit('progress', `setting up the static files bundler`)
-    // 5. inject new action urls into UI
-    await utils.writeConfig(newConfig.web.injectedConfig, newConfig.actions.urls)
-
     // 6. start UI dev server
     const app = express()
     app.use(express.json())
@@ -110,17 +117,22 @@ class ActionServer extends CNAScript {
       logLevel: 1
     })
     app.use(bundler.middleware())
+
+    // start server
     const server = app.listen(port)
     this.emit('progress', `local server running at http://localhost:${port}`)
-    // 7. cleanup on SIGINT
+
+    // cleanup on SIGINT
     const cleanup = err => {
       if (err) console.error(err)
-      console.error('Resetting .env')
-      fs.removeSync('.env')
-      fs.moveSync(DOTENV_SAVE, '.env')
-      console.error('Cleaning up resources')
+      if (!this.config.actions.remote) {
+        console.error('Resetting .env')
+        fs.removeSync('.env')
+        fs.moveSync(DOTENV_SAVE, '.env')
+        console.error('Cleaning up resources')
+        owStack.kill()
+      }
       server.close()
-      owStack.kill()
       err ? process.exit(1) : process.exit(0)
     }
     // todo cleanup on kill, exit, unhandled error as well
